chore(lint): enable unconvert in golangci-lint
Enable the unconvert linter, which flags type conversions where the
value already has the target type (e.g. int(x) where x is already
int). A measured pass found 56 findings: 55 in the root module and 1
in the integration module; the other two modules had none.

Every finding was a genuinely redundant conversion with no behaviour
change, for one of two reasons. Most of the codebase's domain
identifiers (TxNum, BlockNum, Namespace, PKey, TxID, RawValue,
PeerID, PeerIPAddress, LabelName) are declared as true Go type
aliases (`type X = Y`), so converting across them is identical to
converting a type to itself. The remainder - notably FieldName
(`type FieldName string`) and the driver.TransactionType underlying
the TransactionType alias - are genuinely defined types, not
aliases; those deletions are safe for a different reason, namely
that the conversion was a self-conversion of a value already
carrying that exact named type. None of the four watch-for cases
(pinning an untyped constant into an interface, crossing a type
alias/named type, generics, or a deliberate width assertion)
applied, so no nolint was needed - all 56 conversions were deleted
outright.

Removing two of the conversions turned a manual element-by-element
slice copy in Vault.Statuses (fabricx) into two slices of the same
underlying type; replaced the loop with copy() to satisfy
staticcheck's S1001, which the change surfaced.

+++ platform/fabric/core/generic/committer/endorsertx.go
@@ -60,7 +60,7 @@ func MapFinalityEvent(ctx context.Context, block *common.BlockMetadata, txNum dr
 	}
 
 	validationFlags := ValidationFlags(block.Metadata[common.BlockMetadataIndex_TRANSACTIONS_FILTER])
-	if uint64(txNum) >= uint64(len(validationFlags)) {
+	if txNum >= uint64(len(validationFlags)) {
 		return 0, nil, errors.Errorf("transaction number [%d] out of range for validation flags of length [%d]", txNum, len(validationFlags))
 	}
